Move e2e for egressIP into new file under test/e2e
This commit moves the egressIP e2e's to a new file to
avoid cluttering the main e2e.go which is already a
mess in regards to functionality wise split of tests.

(cherry picked from commit 8dafe4df0e2f397d184068efa9a90de1d3a00b26)
(cherry picked from commit bff60e179dcc79df2f871cdd7ce32f5ea6507056)

Conflicts in 4.11:
	test/e2e/e2e.go
because
https://github.com/openshift/ovn-kubernetes/pull/1331/commits/0466bc455c69f090e6940cb302631aa721435891
is missing in 4.11

(cherry picked from commit 5ad3779b0498fdafe4a11495a9a602ab17e13c88)

Conflicts in 4.10:
	test/e2e/e2e.go
